+++ src/main/java/teamport/aether/entity/monster/fireminion/MobFireMinion.java
@@ -8,7 +8,7 @@ import net.minecraft.core.util.helper.DamageType;
 import net.minecraft.core.world.World;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import teamport.aether.ITranslatableDeathMessage;
+import teamport.aether.entity.ITranslatableDeathMessage;
 
 public class MobFireMinion extends MobMonster implements Enemy, ITranslatableDeathMessage {
     public MobFireMinion(@Nullable World world) {
