+++ src/main/java/teamport/aether/entity/monster/sentry/MobSentry.java
@@ -11,7 +11,7 @@ import net.minecraft.core.util.collection.NamespaceID;
 import net.minecraft.core.util.helper.DamageType;
 import net.minecraft.core.world.World;
 import org.jetbrains.annotations.NotNull;
-import teamport.aether.ITranslatableDeathMessage;
+import teamport.aether.entity.ITranslatableDeathMessage;
 import teamport.aether.blocks.AetherBlocks;
 import teamport.aether.items.itemtool.ItemToolPickaxeAether;
