+++ src/main/java/teamport/aether/entity/monster/valkyrie/MobValkyrie.java
@@ -13,7 +13,7 @@ import net.minecraft.core.util.helper.DamageType;
 import net.minecraft.core.util.helper.MathHelper;
 import net.minecraft.core.world.World;
 import org.jetbrains.annotations.NotNull;
-import teamport.aether.ITranslatableDeathMessage;
+import teamport.aether.entity.ITranslatableDeathMessage;
 import teamport.aether.items.AetherItems;
 
 public class MobValkyrie extends MobPathfinder implements Enemy, ITranslatableDeathMessage {
