+++ src/main/java/teamport/aether/entity/projectile/ProjectileArrowFlaming.java
@@ -18,8 +18,9 @@ import net.minecraft.core.util.phys.HitResult;
 import net.minecraft.core.util.phys.Vec3;
 import net.minecraft.core.world.World;
 import org.jetbrains.annotations.Nullable;
+import teamport.aether.entity.ITranslatableDeathMessage;
 
-public class ProjectileArrowFlaming extends ProjectileArrow implements ProjectileAether {
+public class ProjectileArrowFlaming extends ProjectileArrow implements ProjectileAether, ITranslatableDeathMessage {
 
     public ProjectileArrowFlaming(World world) {
         super(world);
